Python/exercise/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
pandas==2.1.1
openpyxl==3.1.2
werkzeug==2.3.7
PyJWT==2.8.0